+++ src/constants.ts
@@ -1,2 +1,1 @@
-export const NEON_DEFAULT_ROLE_NAME = 'neondb_owner';
 export const NEON_DEFAULT_DATABASE_NAME = 'neondb';